fix #5 : finish the unitary testing
need to finish the comportment testing

+++ src/main/java/fr/unice/polytech/cod/order/Customer.java
@@ -22,7 +22,7 @@ public class Customer {
         return true;
     }
 
-    public boolean payBack(int transactionNumber){
+    public boolean payBack(double transactionNumber){
         System.out.print(transactionNumber); //TODO figure out what to do
         return true;
     }
